+++ seo/all.php
@@ -1,6 +1,7 @@
 <?php
 if(is_admin() ) {
   require_once 'h-metabox.php';
+  require_once 'h-profile.php';
 } else {
   require_once 'h-metatags.php';
   require_once 'h-microdata.php';